# koan/openapi.yaml
...
components:
  schemas:
    Error:
      properties:
        error:
          properties:
            code:
              type: string
            message:
              type: string
          required:
          - code
          - message
          type: object
      required:
      - error
      type: object
  securitySchemes:
    bearerAuth:
      scheme: bearer
      type: http
info:
  description: Optional, token-authenticated HTTP control layer for the Kōan agent. See docs/operations/rest-api.md.
  title: Kōan REST API
  version: 1.0.0
openapi: 3.1.0
paths:
...
  /v1/missions:
    get:
      description: 'The store is the same source ``GET /v1/status`` counts, so a mission

        queued from Slack/GitHub/dashboard is visible here even when it was never

        recorded in the API sidecar. Rows are grouped by state in

        ``pending``/``in_progress``/``done``/``failed`` order — queue order within

        the live states, most-recent-first within the terminal ones. ``?status``

        filters to one state, ``?project`` to one project, and ``?limit`` caps the

        rows returned *per state* (so a limit with no ``?status`` can return up to

        four times that many rows).'
      operationId: missions_list_missions_route_get
      parameters:
      - description: Restrict results to one mission-store state.
        in: query
        name: status
        required: false
        schema:
          enum:
          - pending
          - in_progress
          - done
          - failed
          type: string
      - description: Restrict results to one project.
        in: query
        name: project
        required: false
        schema:
          type: string
      - description: Cap the rows returned per state. Omit for unlimited; a non-integer or < 1 value is rejected with 422.
        in: query
        name: limit
        required: false
        schema:
          minimum: 1
          type: integer
      responses:
        '200':
          description: Successful response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Missing or malformed bearer token
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid bearer token
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
      summary: List missions from the authoritative mission store.
      tags:
      - missions
      x-koan-mcp: true
      x-koan-mcp-description: Use this to browse or filter queues. Do not poll it for one mission's completion; use `koan_missions_get` with that mission id.
    post:
      description: 'The mission is appended to the pending queue for a later agent cycle.

        Setting `urgent` inserts it at the front instead.'
      operationId: missions_create_mission_post
      requestBody:
        content:
          application/json:
            schema:
              anyOf:
              - properties:
                  command:
                    pattern: \S
                required:
                - command
              - properties:
                  text:
                    pattern: \S
                required:
                - text
              properties:
                command:
                  anyOf:
                  - enum:
                    - /audit
                    - /brief
                    - /ci_check
                    - /digest
                    - /doc
                    - /docs
                    - /explain
                    - /fix
                    - /gh_request
                    - /impl
                    - /implement
                    - /plan
                    - /rb
                    - /re_review
                    - /rebase
                    - /rereview
                    - /review
                    - /rv
                    - /xp
                    type: string
                  - pattern: ^/?(?:audit|brief|ci_check|digest|doc|docs|explain|fix|gh_request|impl|implement|plan|rb|re_review|rebase|rereview|review|rv|xp)(?:\s+[\s\S]+)?$
                    type: string
                  description: Slash-command mission; prefer this for work covered by an existing Kōan skill. Append arguments to the selected verb, for example "/review https://github.com/owner/repo/pull/42". Use text only when no exposed skill covers the work.
                project:
                  description: Optional project name added as a project tag.
                  type: string
                text:
                  description: Free-form mission text for work no exposed skill covers.
                  type: string
                urgent:
                  default: false
                  description: Insert the mission at the front of the pending queue.
                  type: boolean
              type: object
        required: true
      responses:
        '202':
          description: Successful response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Missing or malformed bearer token
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid bearer token
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
      summary: Queue a new mission.
      tags:
      - missions
      x-koan-mcp: true
      x-koan-mcp-description: Prefer `command` for anything an existing Kōan skill does. Pass the slash command with its arguments, for example "/review https://github.com/owner/repo/pull/42". Call `koan_skills_list` for the full catalogue, usage, aliases, and per-command flags. Use `text` only for work no exposed skill covers. The returned id identifies queued work, not a completed result. Poll `koan_missions_get`, then call `koan_missions_result` after status becomes `done`.
  /v1/missions/reorder:
    post:
      description: Reordering changes queue priority without editing the mission text.
      operationId: missions_reorder_mission_route_post
      requestBody:
        content:
          application/json:
            schema:
              properties:
                mission_id:
                  description: Identifier of the pending mission to move.
                  pattern: \S
                  type: string
                target_position:
                  description: One-indexed position in the pending queue.
                  minimum: 1
                  type: integer
              required:
              - mission_id
              - target_position
              type: object
        required: true
      responses:
        '200':
          description: Successful response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Missing or malformed bearer token
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid bearer token
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
      summary: Move a pending mission to a new position.
      tags:
      - missions
      x-koan-mcp: true
      x-koan-mcp-description: Only pending missions can be reordered. Obtain the mission id from `koan_missions_list` and use a one-indexed target position.
...
  /v1/missions/{mission_id}/result:
    get:
      description: 'Inline and spilled results use the same HTTP response, so clients never

        need filesystem access.'
      operationId: missions_get_mission_result_route_get
      parameters:
      - description: Mission identifier returned when the mission was queued.
        in: path
        name: mission_id
        required: true
        schema:
          type: string
      responses:
        '200':
          description: Successful response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Missing or malformed bearer token
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid bearer token
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
      summary: Fetch a finished mission's result.
      tags:
      - missions
      x-koan-mcp: true
      x-koan-mcp-description: Call this after `koan_missions_get` reports `done`. It returns the complete structured result and reports not found when no structured result is available.
...
  /v1/status:
    get:
      description: 'Combines execution state, queue counts, signal flags, pause details, and

        attention count in one response.'
      operationId: status_status_get
      responses:
        '200':
          description: Successful response
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Missing or malformed bearer token
        '403':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Invalid bearer token
        default:
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
          description: Error response
      summary: Get current agent state, execution and mission counters.
      tags:
      - status
      x-koan-mcp: true
      x-koan-mcp-description: Use this inexpensive call first to orient yourself before choosing a mission, observability, or control tool.
...
security:
- bearerAuth: []
servers:
- description: Default loopback bind
  url: http://127.0.0.1:8420
...